Move timer display into its own component
